width fixed , aded plus, hash

+++ ft_printf.h
@@ -56,7 +56,7 @@ typedef struct	s_data
 	int			pos; //curr pos in format
 	t_flags		info;
 	char 		*arg_string;
-	int 		sym;
+	size_t 		sym;
 }				t_data;
 
 size_t	ft_strlen(const char *s);
